+++ docs/changelog.rst
@@ -2,11 +2,17 @@
 Changelog
 =========
 
+
 * :feature:`-` Page for inputting external API keys has been revamped.
 * :feature:`-` Added a confirmation dialog that appears when users attempt to discard changes on the form dialog, preventing accidental removal of changes.
 * :feature:`8618` rotki will now support Morpho vaults on Ethereum and Base, with related historical events properly decoded and correct prices shown for vault tokens.
 * :feature:`8620` rotki will now properly decode EFP (Ethereum Follow Protocol) transactions.
 * :feature:`8602` rotki will now properly decode Basenames transactions and add calendar reminders for name expiry.
+* :bug:`8916` Coinbase events (trades, deposits/withdrawals, earn etc.) will now be properly pulled after the initial sync. 
+* :bug:`-` Fix the issue where some asset values show zero in the edit snapshot form.
+* :bug:`-` Fix the issue where the pagination for the account table resets to the first page when the user expands the account.
+* :bug:`8892` rotki will now correctly fetch Starknet token prices before May 2024 from Cryptocompare, when the ticker changed from STARK to STRK.
+* :bug:`-` The airdrops directory should no longer appear under the user directory in certain cirsumstances
 
 * :release:`1.36.0 <2024-11-06>`
 * :bug:`-` The exported file that overrides the file with the same name should have the latest modified time.